BED-2549: ensure that the `predis` key is defined in your `cacheOptions` array before it is accessed https://code.go1.com.au/util/app/-/merge_requests/17 https://code.go1.com.au/util/app/-/merge_requests/20
The recent MR regarding `predis` changed the structure of the configuration in `config.default.php`. The current implementation catches the exception and returns `null`, leading to the application doesn't know where the error is.
So If the key does not exist, an exception will be thrown, preventing the undefined array key error. If the key exists but the connection fails, the `try-catch` block will handle the exception.